requirements.txt
anthropic==0.40.0
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
python-dotenv==1.0.1
requests==2.32.3
schedule==1.2.0